feat: Enhance Next.js configuration and improve component structure
- Updated next.config.js to include performance optimizations, image configuration, security headers, and compression settings.
- Modified tsconfig.json to target ES2017 and added strict type-checking options for better code quality.
- Improved layout and metadata in layout.tsx, including updated site URL and enhanced Open Graph and Twitter card data.
- Refactored page.tsx to handle dynamic content loading with error handling and improved user feedback.
- Updated miembros components to reflect new structure and data sources, enhancing maintainability and clarity.
- Removed obsolete miembros-generales component to streamline the codebase.
