+++ next.config.js
@@ -1,4 +1,50 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {};
+const nextConfig = {
+  // Optimizaciones de rendimiento
+  experimental: {
+    optimizePackageImports: ["@heroui/button", "@heroui/navbar", "@heroui/input"],
+  },
+
+  // Configuración de imágenes
+  images: {
+    formats: ["image/webp", "image/avif"],
+    deviceSizes: [640, 750, 828, 1080, 1200, 1920, 2048, 3840],
+    imageSizes: [16, 32, 48, 64, 96, 128, 256, 384],
+    minimumCacheTTL: 60,
+  },
+
+  // Headers de seguridad
+  async headers() {
+    return [
+      {
+        source: "/(.*)",
+        headers: [
+          {
+            key: "X-Frame-Options",
+            value: "DENY",
+          },
+          {
+            key: "X-Content-Type-Options",
+            value: "nosniff",
+          },
+          {
+            key: "Referrer-Policy",
+            value: "origin-when-cross-origin",
+          },
+        ],
+      },
+    ];
+  },
+
+  // Configuración de compresión
+  compress: true,
+
+  // Configuración de PWA (opcional)
+  // pwa: {
+  //   dest: 'public',
+  //   register: true,
+  //   skipWaiting: true,
+  // },
+};
 
 module.exports = nextConfig;
