+++ src/app/layout.tsx
@@ -1,5 +1,5 @@
 import "@/styles/globals.css";
-import { Metadata } from "next";
+import { Metadata, Viewport } from "next";
 import clsx from "clsx";
 
 import { Providers } from "./providers";
@@ -10,20 +10,53 @@ import { Footer } from "@/components/layout/Footer";
 import { Navbar } from "@/components/layout/Navbar";
 
 export const metadata: Metadata = {
-  metadataBase: new URL("https://jciambato.com"),
-  title: "JCI Ambato — Formación de Líderes Juveniles con Impacto Social",
+  metadataBase: new URL("https://jciambato.org"),
+  title: {
+    default: "JCI Ambato — Formación de Líderes Juveniles con Impacto Social",
+    template: "%s | JCI Ambato"
+  },
   description: siteConfig.description,
   keywords: siteConfig.keywords,
   authors: siteConfig.authors,
-  openGraph: siteConfig.openGraph,
+  openGraph: {
+    ...siteConfig.openGraph,
+    type: "website",
+    locale: "es_EC",
+    siteName: "JCI Ambato",
+  },
+  twitter: {
+    card: "summary_large_image",
+    title: "JCI Ambato — Formación de Líderes Juveniles con Impacto Social",
+    description: siteConfig.description,
+    images: ["/images/logos/jci-ambato.webp"],
+  },
   alternates: siteConfig.alternates,
   robots: {
     index: true,
     follow: true,
+    googleBot: {
+      index: true,
+      follow: true,
+      "max-video-preview": -1,
+      "max-image-preview": "large",
+      "max-snippet": -1,
+    },
   },
   icons: {
     icon: "/favicon.ico",
+    shortcut: "/favicon.ico",
+    apple: "/favicon.ico",
   },
+  manifest: "/manifest.json",
+  verification: {
+    google: "your-google-verification-code",
+  },
+};
+
+export const viewport: Viewport = {
+  width: "device-width",
+  initialScale: 1,
+  maximumScale: 1,
 };
 
 export default function RootLayout({ children }: { children: React.ReactNode }) {
