+++ src/app/providers.tsx
@@ -17,5 +17,11 @@ declare module "@react-types/shared" {
 export function Providers({ children }: ProvidersProps) {
   const router = useRouter();
 
-  return <HeroUIProvider navigate={router.push}>{children}</HeroUIProvider>;
+  return (
+    <HeroUIProvider
+      navigate={router.push}
+    >
+      {children}
+    </HeroUIProvider>
+  );
 }
