+++ src/components/sections/miembros/index.ts
@@ -1,4 +1,4 @@
 export { default as Bienvenida } from "./bienvenida";
-export { default as Miembros } from "./miembros";
-export { default as MiembrosGenerales } from "./miembros-generales";
+export { default as Miembros } from "./junta-directiva";
+export { default as MiembrosGenerales } from "./miembros";
 export { default as Senadores } from "./senadores";
