+++ src/data/junta-directiva.ts
@@ -1,0 +1,56 @@
+export const juntaDirectivaData = [
+  {
+    nombre: "Israel Teneda",
+    cargo: "Pasado Presidente",
+    imagen: "/images/miembros/miembro.png",
+    social: { instagram: "#", linkedin: "#" },
+  },
+  {
+    nombre: "Gabriela González",
+    cargo: "Presidente Local",
+    imagen: "/images/miembros/miembro.png",
+    social: { instagram: "#", linkedin: "#" },
+  },
+  {
+    nombre: "Mauricio",
+    cargo: "Director del Senado",
+    imagen: "/images/miembros/miembro.png",
+    social: { instagram: "#", linkedin: "#" },
+  },
+  {
+    nombre: "Romina Benítez",
+    cargo: "Secretaria Local",
+    imagen: "/images/miembros/miembro.png",
+    social: { instagram: "#", linkedin: "#" },
+  },
+  {
+    nombre: "Emilia Santamaría",
+    cargo: "Vicepresidenta Ejecutiva",
+    imagen: "/images/miembros/miembro.png",
+    social: { instagram: "#", linkedin: "#" },
+  },
+  {
+    nombre: "Gabriela González",
+    cargo: "Tesorero Local",
+    imagen: "/images/miembros/miembro.png",
+    social: { instagram: "#", linkedin: "#" },
+  },
+  {
+    nombre: "Miembro Ejemplo 1",
+    cargo: "Vocal",
+    imagen: null,
+    social: { instagram: "#", linkedin: "#" },
+  },
+  {
+    nombre: "Miembro Ejemplo 2",
+    cargo: "Vocal",
+    imagen: null,
+    social: { instagram: "#", linkedin: "#" },
+  },
+  {
+    nombre: "Miembro Ejemplo 3",
+    cargo: "Vocal",
+    imagen: null,
+    social: { instagram: "#", linkedin: "#" },
+  },
+];
