+++ src/data/senadores.ts
@@ -1,0 +1,62 @@
+export const senadoresData = [
+  {
+    nombre: "Scott Belsky",
+    cargo: "Board Member",
+    imagen: "/images/miembros/miembro.png",
+    social: { instagram: "#", linkedin: "#" },
+  },
+  {
+    nombre: "Shona Brown",
+    cargo: "Chair of the Board",
+    imagen: "/images/miembros/miembro.png",
+    social: { instagram: "#", linkedin: "#" },
+  },
+  {
+    nombre: "Mike Cannon-Brookes",
+    cargo: "Board Member",
+    imagen: "/images/miembros/miembro.png",
+    social: { instagram: "#", linkedin: "#" },
+  },
+  {
+    nombre: "Scott Farquhar",
+    cargo: "Board Member",
+    imagen: "/images/miembros/miembro.png",
+    social: { instagram: "#", linkedin: "#" },
+  },
+  {
+    nombre: "Heather M. Fernandez",
+    cargo: "Board Member",
+    imagen: "/images/miembros/miembro.png",
+    social: { instagram: "#", linkedin: "#" },
+  },
+  {
+    nombre: "Sasan Goodarzi",
+    cargo: "Board Member",
+    imagen: "/images/miembros/miembro.png",
+    social: { instagram: "#", linkedin: "#" },
+  },
+  {
+    nombre: "Karen Dykstra",
+    cargo: "Board Member",
+    imagen: "/images/miembros/miembro.png",
+    social: { instagram: "#", linkedin: "#" },
+  },
+  {
+    nombre: "Christian Smith",
+    cargo: "Board Member",
+    imagen: "/images/miembros/miembro.png",
+    social: { instagram: "#", linkedin: "#" },
+  },
+  {
+    nombre: "Steve Sordello",
+    cargo: "Board Member",
+    imagen: "/images/miembros/miembro.png",
+    social: { instagram: "#", linkedin: "#" },
+  },
+  {
+    nombre: "Rich Wong",
+    cargo: "Board Member",
+    imagen: "/images/miembros/miembro.png",
+    social: { instagram: "#", linkedin: "#" },
+  },
+];
